bugfix: Fix various string utilities to work on non-UTF-8 platforms
  or: Get `docker run --rm -v $PWD:/project -w /project signal-android ./gradlew qa` to pass

I'm not sure if there are any non-UTF-8 locales on Android, but the
default locale in the `./reproducible-builds/Dockerfile` image is set
to "C" (ASCII), not a UTF-8 locale.

So adjust StreamUtil and StringUtil to work on platforms with a
non-UTF-8 default locale, so that `./gradlew qa` works inside of the
reproducible-builds image.